workers/requirements.txt
pika==1.3.2
pydantic==1.10.5
python-dotenv==0.21.1
aiosmtplib==2.0.2
backoff==2.2.1